requirements.txt
aiodns==3.0.0
aiohappyeyeballs==2.6.1
aiohttp==3.8.5
aiosignal==1.3.2
anyio==4.9.0
async-timeout==4.0.3
asyncpg==0.28.0
attrs==25.3.0
bcrypt==4.0.1
Brotli==1.1.0
cchardet==2.1.7
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
clerk-sdk-python==0.1.0
click==8.1.8
coverage==7.8.2
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.95.2
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httpx==0.27.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.4.4
nodeenv==1.9.1
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
prisma==0.15.0
propcache==0.3.1
pyasn1==0.6.1
pycares==4.8.0
pycparser==2.22
pydantic==2.7.0
pydantic-core==2.16.2
pydantic-settings==2.9.1
PyJWT==2.8.0
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-cov==6.1.1
pytest-env==1.1.5
pytest-mock==3.14.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
StrEnum==0.4.15
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.13.2
uvicorn==0.22.0
yarl==1.20.0
